requirements.txt
Pillow==10.0.1
Pillow==10.4.0
streamlit==1.37.1
